daemon: Move files into own go package
Moves the Go files of the cilium-agent daemon into its own Go package.
This unifies the structure of the cilium-agent sources with other
command-line interfaces in the source, i.e. cilium, cilium-health and
bugtool already use this structure.

In addition, this also allows downstream consumers to import the daemon
via Go modules.
