+++ daemon/main.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2019 Authors of Cilium
+// Copyright 2016-2020 Authors of Cilium
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,6 +14,10 @@
 
 package main
 
+import (
+	"github.com/cilium/cilium/daemon/cmd"
+)
+
 func main() {
-	daemonMain()
+	cmd.Execute()
 }
